+++ tests/functional/test__big_c.py
@@ -22,6 +22,7 @@ class Test_C(unittest.FunctionalTestCase):
 
     def setUp(self):
         super().setUp()
+        self.settings().set('vintageous_use_sys_clipboard', False)
         self.resetRegisters()
 
     def test_C(self):