+++ tests/functional/test_x.py
@@ -22,6 +22,7 @@ class Test_x(unittest.FunctionalTestCase):
 
     def setUp(self):
         super().setUp()
+        self.settings().set('vintageous_use_sys_clipboard', False)
         self.resetRegisters()
 
     def test_x(self):