+++ tests/nv/vi/test_registers.py
@@ -89,8 +89,9 @@ class RegistersTestCase(unittest.ViewTestCase):
 
     def setUp(self):
         super().setUp()
-        self.view.settings().erase('vintage')
-        self.view.settings().erase('vintageous_use_sys_clipboard')
+        self.settings().erase('vintage')
+        self.settings().erase('vintageous_use_sys_clipboard')
+        self.settings().set('vintageous_use_sys_clipboard', False)
         set_clipboard('')
         registers._reset_data()
         self.registers = Registers()
@@ -100,6 +101,7 @@ class RegistersTestCase(unittest.ViewTestCase):
     def tearDown(self):
         super().tearDown()
         registers._reset_data()
+        self.resetRegisters()
 
     def assertEmptyRegisters(self):
         self.assertEqual(registers._data, {'0': None, '1-9': deque([None] * 9, maxlen=9)})